requirements.txt
django==4.2.2
python-dotenv
langchain
langchain-community
langchain-openai
duckduckgo-search
